// src/models/unit/Unit.ts
import logo from './pz2.png'
import { Hex } from "../Hex";
import { Fraction } from '../Fraction';




export enum UnitNames {
    UNIT = "Юнит",
    PZ2C = "Легкий танк Pz2C",
    INFANTRYWERNAHT = "Немецкая пехота",
    GUN45SOV = "53-К Пушка",
}

export class Unit {
    fraction: Fraction;
    logo: string;
    hex: Hex;
    name: UnitNames;
    id: number;
    armor: number;
    strength: number;
    ammo: number;



    constructor(fraction: Fraction, hex: Hex, armor: number, strength: number) {
        this.fraction = fraction;
        this.hex = hex;
        this.hex.units = [];
        this.logo = logo;
        this.name = UnitNames.UNIT
        this.id = Math.random()
        this.ammo = 10;
        this.armor = armor;
        this.strength = strength;
    }

}